Renames the main file to search.
